now detecing danlging cnames again!Refactor DNS handling, add dnspython, and improve logging
Replaced aiodns with dnspython for DNS resolution to enhance error handling. Introduced EvidenceCollector class to modularize evidence collection functionality. Added extensive logging to improve debugging and monitoring of DNS resolution processes.
